==> notebooks/07_eqtls/2_subset.ipynb <==
import pandas as pd
import anndata

eqtl_file = 'susie_backmerged_dedup.csv'
eqtl_ad_file = 'eqtl.h5ad'

eqtl_df = pd.read_csv(eqtl_file)
eqtl_ad = anndata.read_h5ad(eqtl_ad_file)

## Subset to high PIP and negatives

eqtl_df = eqtl_df[(eqtl_df.pip > 0.9) | (eqtl_df.cs_id == "negative")]
eqtl_df['label'] = eqtl_df.pip > 0.9

print(len(eqtl_df), eqtl_df['label'].sum(), (~eqtl_df['label']).sum())

eqtl_ad.obs['index'] = eqtl_ad.obs.index.tolist()

keep = eqtl_ad.obs.merge(eqtl_df, on=['gene_id', 'variant'], how='inner')['index_x'].unique()

eqtl_ad = eqtl_ad[keep, :]

eqtl_ad.shape

## Save

eqtl_df.to_csv('scored_sceqtls.csv')
eqtl_ad.write_h5ad('scored_sceqtl_vars.h5ad')
